+++ clang-tools-extra/clang-tidy/hooks/HooksTidyModule.cpp
@@ -10,6 +10,7 @@
 #include "../ClangTidyModule.h"
 #include "../ClangTidyModuleRegistry.h"
 #include "AccountBufLenCheck.h"
+#include "AccountConvBufLenCheck.h"
 #include "EntryPointRecursionCheck.h"
 #include "EntryPointsCheck.h"
 #include "EntryPointsNegCheck.h"
@@ -29,6 +30,8 @@ public:
   void addCheckFactories(ClangTidyCheckFactories &CheckFactories) override {
     CheckFactories.registerCheck<AccountBufLenCheck>(
         "hooks-account-buf-len");
+    CheckFactories.registerCheck<AccountConvBufLenCheck>(
+        "hooks-account-conv-buf-len");
     CheckFactories.registerCheck<EntryPointRecursionCheck>(
         "hooks-entry-point-recursion");
     CheckFactories.registerCheck<EntryPointsCheck>(